Initial move of shared functions to BaseSock
