+++ StreamClient/ActiveSock.cpp
@@ -17,18 +17,6 @@ static char THIS_FILE[] = __FILE__;
 CActiveSock::CActiveSock(HANDLE StopEvent)
   : CBaseSock(StopEvent)
 {
-	//
-	// Initialize the WinSock subsystem.
-	//
-	WSADATA wsadata;
-
-	if (WSAStartup(MAKEWORD(1,1), &wsadata) == SOCKET_ERROR)
-	{
-		DebugMsg("Error %d returned by WSAStartup", GetLastError());
-		throw "WSAStartup error";
-	}
-	int rc = true;
-	setsockopt(ActualSocket, IPPROTO_TCP, TCP_NODELAY, (char *)&rc, sizeof(int));
 }
 
 CActiveSock::~CActiveSock()
@@ -118,115 +106,14 @@ bool CActiveSock::Connect(LPCTSTR HostName, USHORT PortNumber)
 	//	return false;
 	//}
 
- 	if (!read_event)
- 	{
-		read_event = WSACreateEvent();
- 	}
-
-	if (read_event != WSA_INVALID_EVENT)
-	{
-		if (!write_event)
-		{
-			write_event = WSACreateEvent();
-		}
-		if (write_event != WSA_INVALID_EVENT)
-		{
-			if (WSAResetEvent(read_event))
-			{
-				if (WSAResetEvent(write_event))
-				{
-					return true;
-				}
-			}
-		}
-	}
-
-	LastError = WSAGetLastError();
-	return false;
+	return SUCCEEDED(Setup());
 }
 
-// Receives up to Len bytes of data and returns the amount received - or SOCKET_ERROR if it times out
-int CActiveSock::RecvPartial(LPVOID lpBuf, const size_t Len)
-{
-	DWORD bytes_read = 0;
-	DWORD msg_flags = 0;
-
-	// Setup up the events to wait on
-	WSAEVENT hEvents[2] = { m_hStopEvent, read_event };
-	// Create the overlapped I/O event and structures
-	memset(&os, 0, sizeof(OVERLAPPED));
-	os.hEvent = hEvents[1];
-	if (!WSAResetEvent(os.hEvent))
-	{
-		LastError = WSAGetLastError();
-		return SOCKET_ERROR;
-	}
-
-	// Setup the buffers array
-	WSABUF buffer{ static_cast<ULONG>(Len), static_cast<char*>(lpBuf) };
-
-	// If the timer has been invalidated, restart it
-	const auto RecvEndTime = CTime::GetCurrentTime() + CTimeSpan(0, 0, 0, RecvTimeoutSeconds);
-
-	const int rc = WSARecv(ActualSocket, &buffer, 1, &bytes_read, &msg_flags, &os, nullptr); // Start an asynchronous read
-	LastError = WSAGetLastError();
-
-	const CTimeSpan TimeLeft = RecvEndTime - CTime::GetCurrentTime();
-	const auto SecondsLeft = TimeLeft.GetTotalSeconds();
-	if (SecondsLeft <= 0)
-	{
-		LastError = ERROR_TIMEOUT;
-		return SOCKET_ERROR;
-	}
-
-	bool IOCompleted = !rc; // if rc is zero, the read was completed immediately
-
-	// Now wait for the I/O to complete if necessary, and see what happened
-
-	if ((rc == SOCKET_ERROR) && (LastError == WSA_IO_PENDING))  // Read in progress, normal case
-	{
-		const DWORD dwWait = WaitForMultipleObjects(2, hEvents, false, static_cast<DWORD>(SecondsLeft) * 1000);
-		switch (dwWait)
-		{
-		case WAIT_OBJECT_0 + 1: // The read event 
-			IOCompleted = true;
-			LastError = 0;
-			break;
-		case WAIT_ABANDONED_0:
-		case WAIT_ABANDONED_0 + 1:
-			break;
-		case WAIT_TIMEOUT:
-			LastError = ERROR_TIMEOUT;
-			break;
-		case WAIT_FAILED:
-			LastError = ::GetLastError();
-			break;
-		default:
-			break;
-		}
-	}
-
-	if (IOCompleted)
-	{
-		if (WSAGetOverlappedResult(ActualSocket, &os, &bytes_read, true, &msg_flags) && (bytes_read > 0))
-		{
-			return bytes_read; // Normal case, we read some bytes, it's all good
-		}
-		else
-		{	// A bad thing happened
-			const int e = WSAGetLastError();
-			if (e == 0) // The socket was closed
-				return 0;
-			else if (LastError == 0)
-				LastError = e;
-		}
-	}
-	return SOCKET_ERROR;
-}
 
 // Receives exactly Len bytes of data and returns the amount received - or SOCKET_ERROR if it times out
 int CActiveSock::RecvMsg(LPVOID lpBuf, const size_t Len)
 {
+	StartRecvTimer();
 	size_t total_bytes_received = 0;
 	while (total_bytes_received < Len)
 	{
@@ -241,139 +128,10 @@ int CActiveSock::RecvMsg(LPVOID lpBuf, const size_t Len)
 	return (static_cast<int>(total_bytes_received));
 }
 
-void CActiveSock::SetRecvTimeoutSeconds(int NewRecvTimeoutSeconds)
-{
-	if (NewRecvTimeoutSeconds == INFINITE)
-		NewRecvTimeoutSeconds = MAXINT;
-	if (NewRecvTimeoutSeconds > 0)
-	{
-		RecvTimeoutSeconds = NewRecvTimeoutSeconds;
-		// RecvEndTime is untouched because a receive may be in process
-	}
-}
-
-int CActiveSock::GetRecvTimeoutSeconds() const
-{
-	return RecvTimeoutSeconds;
-}
-
-void CActiveSock::SetSendTimeoutSeconds(int NewSendTimeoutSeconds)
-{
-	if (NewSendTimeoutSeconds == INFINITE)
-		NewSendTimeoutSeconds = MAXINT;
-	if (NewSendTimeoutSeconds > 0)
-	{
-		SendTimeoutSeconds = NewSendTimeoutSeconds;
-		// SendEndTime is untouched, because a Send may be in process
-	}
-}
-
-int CActiveSock::GetSendTimeoutSeconds() const
-{
-	return SendTimeoutSeconds;
-}
-
-HRESULT CActiveSock::Disconnect()
-{
-	LastError = ERROR_SUCCESS;
-
-	if (ActualSocket == INVALID_SOCKET)
-		LastError = ERROR_HANDLES_CLOSED;
-	else if (WSACloseEvent(read_event) && WSACloseEvent(write_event) && CloseAndInvalidateSocket() && (WSACleanup() == 0))
-		DebugMsg("Disconnect succeeded");
-	else
-	{
-		LastError = ::WSAGetLastError();
-		DebugMsg("Disconnect failed, WSAGetLastError returned 0x%.8x", LastError);
-	}
-	return HRESULT_FROM_WIN32(LastError);
-}
-
-//sends a message, or part of one
-int CActiveSock::SendPartial(LPCVOID lpBuf, const size_t Len)
-{
-	DebugMsg("CActiveSock::SendPartial, Len = %d", Len);
-	
-	DWORD bytes_sent = 0;
-
-	// Setup up the events to wait on
-	WSAEVENT hEvents[2] = { m_hStopEvent, write_event };
-
-	// Create the overlapped I/O event and structures
-	memset(&os, 0, sizeof(OVERLAPPED));
-	os.hEvent = hEvents[1];
-	if (!WSAResetEvent(os.hEvent))
-	{
-		LastError = WSAGetLastError();
-		return SOCKET_ERROR;
-	}
-
-	// Setup the buffer array
-	WSABUF buffer{ static_cast<ULONG>(Len), static_cast<char*>(const_cast<void*>(lpBuf)) };
-
-	// If the timer has been invalidated, restart it
-	const auto SendEndTime = CTime::GetCurrentTime() + CTimeSpan(0, 0, 0, SendTimeoutSeconds);
-
-	const int rc = WSASend(ActualSocket, &buffer, 1, &bytes_sent, 0, &os, nullptr);
-	LastError = WSAGetLastError();
-
-	const CTimeSpan TimeLeft = SendEndTime - CTime::GetCurrentTime();
-	const auto SecondsLeft = TimeLeft.GetTotalSeconds();
-	if (SecondsLeft <= 0)
-	{
-		LastError = ERROR_TIMEOUT;
-		return SOCKET_ERROR;
-	}
-
-	bool IOCompleted = !rc; // if rc is zero, the write was completed immediately, which is common
-
-	// Now wait for the I/O to complete if necessary, and see what happened
-
-	if ((rc == SOCKET_ERROR) && (LastError == WSA_IO_PENDING))  // Write in progress
-	{
-		const DWORD dwWait = WaitForMultipleObjects(2, hEvents, false, static_cast<DWORD>(SecondsLeft) * 1000);
-		switch (dwWait)
-		{
-		case WAIT_OBJECT_0 + 1: // The write event
-			IOCompleted = true;
-			LastError = 0;
-			break;
-		case WAIT_ABANDONED_0:
-		case WAIT_ABANDONED_0 + 1:
-			break;
-		case WAIT_TIMEOUT:
-			LastError = ERROR_TIMEOUT;
-			break;
-		case WAIT_FAILED:
-			LastError = ::GetLastError();
-			break;
-		default:
-			break;
-		}
-	}
-
-	if (IOCompleted)
-	{
-		DWORD msg_flags = 0;
-		if (WSAGetOverlappedResult(ActualSocket, &os, &bytes_sent, true, &msg_flags))
-		{
-			return bytes_sent;
-		}
-		else
-		{	// A bad thing happened
-			const int e = WSAGetLastError();
-			if (e == 0) // The socket was closed
-				return 0;
-			else if (LastError == 0)
-				LastError = e;
-		}
-	}
-	return SOCKET_ERROR;
-}
-
 //sends all the data or returns a timeout
 int CActiveSock::SendMsg(LPCVOID lpBuf, const size_t Len)
 {
+	StartSendTimer();
 	ULONG total_bytes_sent = 0;
 	while (total_bytes_sent < Len)
 	{
@@ -389,12 +147,4 @@ int CActiveSock::SendMsg(LPCVOID lpBuf, const size_t Len)
 			total_bytes_sent += bytes_sent;
 	}; // loop
 	return (total_bytes_sent);
-}
-
-// Retrrn true if it worked
-bool CActiveSock::CloseAndInvalidateSocket()
-{
-	const auto nRet = closesocket(ActualSocket);
-	ActualSocket = INVALID_SOCKET;
-	return nRet == 0;
 }
