+++ StreamServer/PassiveSock.h
@@ -11,14 +11,14 @@ class CPassiveSock : private CBaseSock, public ISocketStream
 public:
 	explicit CPassiveSock(SOCKET, HANDLE);
 	virtual ~CPassiveSock();
-	void SetTimeoutSeconds(int NewTimeoutSeconds);
-	void ArmRecvTimer();
-	void ArmSendTimer();
+	using CBaseSock::SetTimeoutSeconds;
+	using CBaseSock::StartRecvTimer;
+	using CBaseSock::StartSendTimer;
 	int ReceiveBytes(void * const lpBuf, const size_t Len);
 	int SendBytes(const void * const lpBuf, const size_t Len);
 	// BOOL ShutDown(int nHow = SD_SEND); // ShutDown is no longer public
-	// ISocketStream interface
-	DWORD GetLastError() const override;
+	// ISocketStream interface, these items have to be defined so a cast to ISocketStream works
+	DWORD GetLastError() const override; // Has to be declared so that a castt to 
 	// Receives exactly Len bytes of data and returns the amount sent - or SOCKET_ERROR if it times out
 	int RecvPartial(LPVOID lpBuf, const size_t Len) override;
 	// Sends up to Len bytes of data and returns the amount sent - or SOCKET_ERROR if it times out
@@ -30,8 +30,5 @@ protected:
 	using CBaseSock::m_hStopEvent;
 
 private:
-	CTime RecvEndTime;
-	CTime SendEndTime;
-	int TimeoutSeconds = 1;
 };
